bashmarks: Updates to source data file only when necessary
The _load_bookmarks caches ENV variable & bookmark names to be used when
jumping into a directory and during completion generation.

The list (l) command loads bookmarks without condition. Effectively
becoming a load command as well - a workaround for cases where changes in
another terminal window are invisible to this one.

Renames the path global variable to _bm_path for consistency and to prevent
possible overriding of other paths.