Introduces the Live Ripgrep Telescope picker
A string grep-er with files filter using ripgrep's type and glob options

It splits Telescope prompt with 2 consecutive spaces - "  ", where the
first part is the string to search for, and the second is considered a
list of globs.
A glob is everything ripgrep accepts as a value for the
`--glob` option or one of the predefined shortcuts - see TYPE_SHORTCUTS
and SHORTCUTS tables. Shortcuts can be negated by prefixing exclamation
mark - !.

E.g.
  `ruby  *.rb` - search for string "ruby" in `*.rb` files
  `ruby  rb` - search for string "ruby" in ruby file types using the rb shortcut
  `important  js css` - search "important" in Js and CSS files
  `ruby  rb !spec` - find "ruby" in Ruby files except RSpec spec files

`live_ripgrep` accepts `search` and `word_match` options similar to the
Telescope's `grep_string` builtin picker.

Now it powers word_under_cursor, selection & live_grep finders.